% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_processing_resolution.R
\name{check_processing_resolution}
\alias{check_processing_resolution}
\title{Check processing resolution of species.}
\usage{
check_processing_resolution(input_species, raster_folder, output_dir)
}
\arguments{
\item{input_species}{Species range of the species for which the processing resolution is calculated. Class sf.}

\item{raster_folder}{Folder that contains bioclim data. eg milkunize("Data_RAW/WorldClim/bioclim")}

\item{output_dir}{Directory in which to store the results. eg. milkunize("Projects/Amphibians/R/Output/Modeling_resolution/")}
}
\value{
A string with the processing resolution
}
\description{
Checks with which processing resolution the species has 30 or more points.
}
\examples{
None.
}
