% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distance_surface.R
\name{create_absence_points}
\alias{create_absence_points}
\title{Create absence points based on distance surface}
\usage{
create_absence_points(species_distance_surface, species_range, return = "df",
  points_number = 10000)
}
\arguments{
\item{species_distance_surface}{Distance surface. Output of \code{get_distance_surface}.}

\item{species_range}{Species range. Should be of class \code{sf}.}

\item{return}{Type of object to return. One of \code{c("df", "sp", "sf")}. Default is "df".}

\item{points_number}{Number of absence points to sample.}
}
\value{
Depends on \code{return} argument.
}
\description{
Create absence points based on distance surface
}
\examples{
None.
}
