% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_species_data.R
\name{create_species_data}
\alias{create_species_data}
\title{Create species data for amphibians}
\usage{
create_species_data(processing_species, species_resolution_list, amphibians_shp)
}
\arguments{
\item{processing_species}{Species name.}

\item{species_resolution_list}{File with species resolution.}

\item{amphibians_shp}{Shapefile with all of amphibians data.}
}
\value{
PA data.
}
\description{
Create species data (presences and absences) based on processing resolution.
}
\examples{
None.
}
