% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_processing_resolution.R
\name{get_cell_number}
\alias{get_cell_number}
\title{Get cell number
Function takes in species range and a raster with
specified resolution, and returns the number of cells
that exist within the species range}
\usage{
get_cell_number(input_species, resolution, dir_path = "WorldClim-1.4/current")
}
\arguments{
\item{input_species}{sp class species spatial data.}

\item{resolution}{Resolution to check. One in c("10m", "5m", "2.5m", "30s").}
}
\value{
Integer
}
\description{
Get cell number
Function takes in species range and a raster with
specified resolution, and returns the number of cells
that exist within the species range
}
\examples{
None

}
