% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distance_surface.R
\name{get_distance_surface}
\alias{get_distance_surface}
\title{Calculate distance surface}
\usage{
get_distance_surface(species_range, output_name, processing_resolution_data,
  keep = FALSE)
}
\arguments{
\item{species_range}{Species range. Should be class \code{sf}.}

\item{output_name}{Output name of created SAGA grid.}

\item{processing_resolution_data}{csv with species names and processing resolutions.}
}
\value{
Raster
}
\description{
Take species range and raster mask, and create distance surface
that will encompass same realm and continent that species itself occupies (absence creation regions).
This function will call SAGA GIS from the terminal, so SAGA has to be installed, and able to
}
\examples{
None.
library(Rahat); library(raster); library(tidyverse)

proc_res <- "Projects/Amphibians/Output/csv/Species_processing_resolution.csv" \%>\%
 milkunize2() \%>\%
 read_csv()


}
