% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_habitat_suitability.R
\name{get_habitat_suitability}
\alias{get_habitat_suitability}
\title{Get habitat suitability}
\usage{
get_habitat_suitability(species_habitat, conversion_matrix, species_name)
}
\arguments{
\item{species_habitat}{Test.}

\item{conversion_matrix}{Test.}

\item{species_name}{Test.}
}
\value{
dataframe
}
\description{
Pass IUCN habitat classification preferences, and habitat classification matrix, and the function
will return a dataframe that contains information whether the specific habitat is suitable or unsuitable.
Currently the function is intended for use in the amphibians project, and some options that a general user
might want to change are hardcoded.
}
\details{
Use \code{rredlist::rl_habitats()} function to retrieve species habitat preference data. Currently only
IUCN -> ESA CCI land cover conversion matrix is supported.
}
\examples{
None.

library(rredlist)
library(Rahat)
library(tidyverse)

habitat_conversion_matrix <- "Projects/Amphibians/Documents/Notes/Habitat_suitability/Habitats_conversion.csv" \%>\%
milkunize2() \%>\%
read_csv() \%>\%
mutate(IUCN_code = as.character(IUCN))

my_sp <- "Salamandra salamandra"
# Get IUCN species habitat data preferences
# Originally stored in individual csv's, here everything is stored in one rds file
species_habitat_data <- "Projects/Amphibians/data/IUCN_habitat_data/Species_habitat_data.rds" \%>\%
  milkunize2() \%>\%
read_rds()

sp_data <- species_habitat_data \%>\%
filter(Species == my_sp)

habitat_data <- rl_habitats(my_sp)$results
get_habitat_suitability(species_habitat = sp_data,
conversion_matrix = habitat_conversion_matrix, species_name = my_sp)
}
