% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distance_surface.R
\name{get_presence_points}
\alias{get_presence_points}
\title{Get presence points}
\usage{
get_presence_points(species_range, processing_resolution_data)
}
\arguments{
\item{species_range}{Species range. Should be of class sf.}

\item{processing_resolution_data}{Processing resolution csv.}
}
\value{
dataframe
}
\description{
Get presence points from species ranges. This function will take species range,
and retrieve processing resolution (resolution at which there are min 30 points),
and create presence points at given resolution. Maximum number of points is 1000.
If there are more than 1000 points, they will be sampled so there is max 1000.
}
\examples{
None
}
