% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_processing_resolution.R
\name{get_processing_resolution}
\alias{get_processing_resolution}
\title{Get processing resolution for species data}
\usage{
get_processing_resolution(species)
}
\arguments{
\item{species}{Species range of the species for which the processing resolution is calculated. Class sp.}
}
\value{
String.
}
\description{
This function will test the get_cell_number function against rasters of 4 different
spatial resolutions. Checks with which processing resolution the species has 30 or more points.
}
\examples{
None.
}
