% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distance_surface.R
\name{load_mask}
\alias{load_mask}
\title{Load realm or continent mask}
\usage{
load_mask(type, resolution, path = "Projects/Amphibians/data_raw/Rasters")
}
\arguments{
\item{type}{Type of data to load. One in c("Realm", "Continent", "Climate")}

\item{resolution}{Data resolution. One in c("10m", "5m", "2.5m", "30s")}

\item{path}{Folder path. Default is milkunize2("Projects/Amphibians/data_raw/Rasters")}
}
\value{
Raster*
}
\description{
Load realm or continent mask
}
\examples{
None
}
